Chore: removed docs file to be auto-generated Chore: moved prompt methods to command file
